Use emscripten directory to store embedded config
This means that each version of emscripten can have its own cache
and config rather than emsdk holding single global one.

One of the main reasons for doing it this way is so that we can
introduce the concept of an embedded config upstream in emscripten.

Eventually I'd like to always use embedded config both with emsdk
and emscripten.  See #9543